[refactor] cleanup and simplify elibility flow.
