+++ src/app/(BottomTabNavigation)/AllCases/ClaimActions/FileClaim/[caseUid].tsx
@@ -1,27 +1,24 @@
 import { router, useLocalSearchParams } from 'expo-router';
 import React, { useEffect, useState } from 'react';
-import { View, Text, TouchableOpacity } from 'react-native';
+import { View, Text } from 'react-native';
 
-import styles from './styles';
-import BlackRightArrow from '../../../../../assets/black-right-arrow.svg';
-import Document from '../../../../../assets/document-add.svg';
-import Checkbox from '../../../../../assets/double-checkbox.svg';
-import Fileclaim from '../../../../../assets/file-claim-small.svg';
-import RightWhiteArrow from '../../../../../assets/right-arrow-white.svg';
+import BlackRightArrow from '../../../../../../assets/black-right-arrow.svg';
+import Document from '../../../../../../assets/document-add.svg';
+import Checkbox from '../../../../../../assets/double-checkbox.svg';
+import Fileclaim from '../../../../../../assets/file-claim-small.svg';
+import RightWhiteArrow from '../../../../../../assets/right-arrow-white.svg';
+import {
+  ButtonBlack,
+  ButtonWhite,
+} from '../../../../../Components/AuthButton/AuthButton';
+import { fonts } from '../../../../../styles/fonts';
 import {
   getCaseById,
   updateCaseStatus,
-} from '../../../../supabase/queries/cases';
-import { Case, CaseUid, Eligibility } from '../../../../types/types';
-import { openUrl } from '../utils';
-
-const ensureURLFormat = (url: string | null | undefined) => {
-  if (!url) return null;
-  if (url.startsWith('http://') || url.startsWith('https://')) {
-    return url;
-  }
-  return `https://${url}`;
-};
+} from '../../../../../supabase/queries/cases';
+import { Case, CaseUid, Eligibility } from '../../../../../types/types';
+import { openUrl } from '../../utils';
+import styles from '../styles';
 
 export default function FileClaimScreen() {
   const { caseUid } = useLocalSearchParams<{ caseUid: CaseUid }>();
@@ -59,34 +56,33 @@ export default function FileClaimScreen() {
       ) : (
         <View style={styles.screenContainer}>
           <View style={styles.contentContainer}>
-            <Text style={styles.titleText}>Filing a claim</Text>
+            <Text style={fonts.tabHeading}>Filing a claim</Text>
+
             <View style={styles.infoContainer}>
-              <View style={styles.textIconContainer}>
-                <Fileclaim style={styles.icon} />
+              <View style={styles.infoRow}>
+                <Fileclaim />
                 <View style={styles.textContainer}>
-                  <Text style={styles.infoText}>
+                  <Text style={fonts.greyBody}>
                     Filing a claim commits you to the class action and any
                     potential compensation awarded
                   </Text>
                 </View>
               </View>
-            </View>
-            <View style={styles.infoContainer}>
-              <View style={styles.textIconContainer}>
-                <Document style={styles.icon} />
+
+              <View style={styles.infoRow}>
+                <Document />
                 <View style={styles.textContainer}>
-                  <Text style={styles.infoText}>
+                  <Text style={fonts.greyBody}>
                     Complete the required claim form on the official claim
                     filing site to submit your claim
                   </Text>
                 </View>
               </View>
-            </View>
-            <View style={styles.infoContainer}>
-              <View style={styles.textIconContainer}>
-                <Checkbox style={styles.icon} />
+
+              <View style={styles.infoRow}>
+                <Checkbox />
                 <View style={styles.textContainer}>
-                  <Text style={styles.infoText}>
+                  <Text style={fonts.greyBody}>
                     Come back to this screen to update your claim filing status
                   </Text>
                 </View>
@@ -95,24 +91,17 @@ export default function FileClaimScreen() {
           </View>
 
           <View style={styles.buttonsContainer}>
-            <TouchableOpacity
-              style={[styles.buttonBase, styles.buttonWhite]}
-              onPress={() => navigateToClaimLink()}
-            >
-              <Text style={[styles.buttonText, styles.blackText]}>
+            <ButtonWhite onPress={() => navigateToClaimLink()}>
+              <Text style={fonts.blackButton}>
                 Take me to claim filing site
               </Text>
               <BlackRightArrow />
-            </TouchableOpacity>
-            <TouchableOpacity
-              style={[styles.buttonBase, styles.buttonBlack]}
-              onPress={() => confirmClaimFiled()}
-            >
-              <Text style={[styles.buttonText, styles.whiteText]}>
-                I’ve already filed a claim!
-              </Text>
+            </ButtonWhite>
+
+            <ButtonBlack onPress={() => confirmClaimFiled()}>
+              <Text style={fonts.whiteButton}>I’ve already filed a claim!</Text>
               <RightWhiteArrow />
-            </TouchableOpacity>
+            </ButtonBlack>
           </View>
         </View>
       )}
